Styled existing pages. Merged Settings page into Game page. Added mock data for the player scoreboard. Embedded fonts from Google Fonts: Roboto Mono, IBM Plex Sans
