+++ src/components/game/Board/Board.module.css
@@ -1,4 +1,3 @@
 .board {
     display: flex;
-    margin-top: 1rem;
 }
