+++ src/components/ui/Button/Button.module.css
@@ -9,7 +9,7 @@
     cursor: pointer;
 
     height: 2rem;
-    margin: 0;
+    margin: 0.5rem 0;
     padding: 0 0.875rem;
 
     border: 1px solid var(--bg-4);
@@ -22,5 +22,11 @@
 
     &:active {
         background-color: var(--accent-5);
+    }
+
+    &:disabled {
+        background-color: var(--bg-0);
+        color: var(--text-4);
+        cursor: not-allowed;
     }
 }
