+++ src/components/ui/Formik/Select/Select.module.css
@@ -7,7 +7,7 @@
 .Select select {
     color: var(--text-1);
     font-family: inherit;
-    font-size: 1rem;
+    font-size: 0.9rem;
     font-weight: 400;
     line-height: 1.5rem;
