+++ src/components/ui/Title/Title.module.css
@@ -1,10 +1,10 @@
 .title {
     background: repeating-linear-gradient(
         45deg,
-        rgba(128, 128, 128, 0.1),
-        rgba(128, 128, 128, 0.1) 10px,
+        rgba(128, 128, 128, 0.07),
+        rgba(128, 128, 128, 0.07) 10px,
         transparent 10px,
         transparent 20px
     );
-    padding-left: 1rem;
+    padding: 0 1rem;
 }
