+++ src/pages/Game/Game.module.css
@@ -1,0 +1,17 @@
+.Page {
+    width: 100%;
+    align-items: center;
+    display: flex;
+    flex-direction: column;
+}
+
+.ModalContent {
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+}
+
+.ModalButtons {
+    display: flex;
+    gap: 0.5rem;
+}
