+++ src/pages/Results/Results.module.css
@@ -1,0 +1,24 @@
+.Page {
+    width: 100%;
+    align-items: center;
+    display: flex;
+    flex-direction: column;
+}
+
+.Score {
+    font-family: "Roboto Mono", system-ui, Avenir, Helvetica, Arial, sans-serif;
+}
+
+.Score a {
+    display: block;
+    padding: 0 1rem;
+}
+
+.Score a:nth-of-type(2n) {
+    background-color: var(--bg-4);
+}
+
+.Score a:hover {
+    background-color: var(--bg-3);
+    cursor: pointer;
+}
