+++ src/pages/Welcome/Welcome.module.scss
@@ -19,7 +19,7 @@
 .Author {
     opacity: 0;
     font-size: 0.65rem;
-    margin-bottom: 1rem;
+    margin-bottom: 0.5rem;
     animation: fade 1s ease-out forwards;
 }
