+++ src/pages/Welcome/WelcomePage.tsx
@@ -1,5 +1,5 @@
 import style from './Welcome.module.scss';
-import Button from "../components/ui/Button/Button.tsx";
+import Button from "../../components/ui/Button/Button.tsx";
 import {useNavigate} from "react-router-dom";
 
 const WelcomePage = () => {
